% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_twbx.R
\name{update_twbx}
\alias{update_twbx}
\title{Updates a Tableau .twbx file with new data.}
\usage{
update_twbx(files, twbx_file)
}
\arguments{
\item{files}{input character vector}
}
\value{
Nothing
}
\description{
Updates a Tableau .twbx file with new data.
}
\examples{
update_twbx("new_data.xlsx","cool_viz.twbx")
}
